some changes in userService and ImageDeleting

+++ src/main/java/project/service/UserService.java
@@ -1,15 +1,10 @@
 package project.service;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -18,27 +13,31 @@ import project.exception.ResourceNotFoundException;
 import project.model.User;
 import project.repository.UserRepository;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Optional;
 
 @Service
 public class UserService  {
 
     private final UserRepository userRepository;
+    private final ItemService itemService;
 
     private PasswordEncoder passwordEncoder() {
         return new BCryptPasswordEncoder();
 
     }
-    public UserService(UserRepository userRepository) {
+    public UserService(UserRepository userRepository, ItemService itemService) {
 
         this.userRepository = userRepository;
+        this.itemService = itemService;
     }
 
     public User getAuthenticatedUser() {
         Authentication auth = SecurityContextHolder.getContext().getAuthentication();
         org.springframework.security.core.userdetails.User springUser = (org.springframework.security.core.userdetails.User) auth.getPrincipal();
-        User user = userRepository.findByEmail(springUser.getUsername());
-        return user;
+        return userRepository.findByEmail(springUser.getUsername());
     }
     @Transactional
     public void saveUser(User user) {
@@ -79,9 +78,15 @@ public class UserService  {
         }
         userRepository.save(user);
     }
+
     @Transactional
-    public void deleteUser(User user) {
-        userRepository.delete(user);
+    public void deleteUser(String password) {
+        User user = getAuthenticatedUser();
+
+        if (passwordEncoder().matches(password, user.getPassword())) {
+            user.getItemList().forEach(itemService::deleteItem);
+            userRepository.delete(user);
+        }
     }
 
     public User getUserById(Integer id) {
@@ -104,8 +109,8 @@ public class UserService  {
     }
     @Transactional
     public boolean changePassword(String oldPassword, String newPassword) {
-        if (passwordEncoder().matches(oldPassword, newPassword)) {
-            User user = getAuthenticatedUser();
+        User user = getAuthenticatedUser();
+        if (passwordEncoder().matches(oldPassword, user.getPassword())) {
             user.setPassword(passwordEncoder().encode(newPassword));
             userRepository.save(user);
 
@@ -114,9 +119,12 @@ public class UserService  {
 
     }
 
-    public void deleteUser(String password) {
-        User user = getAuthenticatedUser();
+    public void deleteUserFromAdminPanel(int  id) {
+        User user = userRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Item", "id", id));
+        user.getItemList()
+                .forEach(itemService::deleteItem);
+
         userRepository.delete(user);
-        //TODO ARO jnjel bolor 
     }
 }
